Added New GUI
Added a Logout GUI to help with the users logging out of the game. I also created a queue that will start if there are too many people online. When users logout of the game, the queue will pop and then the popped user in the queue can log into the game.