Added View Accounts Ability
Added a view accounts gui that allows the user to see how many accounts are in the game. It shows the users passwords, so in a later update it will be fixed to only show the users emails. Also added a sorting algorithm that sorts the users emails by alphabetical order, so when the user views accounts it should be in alphabetical order.